pyproject.toml
[project]
name = "mag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "neo4j>=6.1.0",
    "python-dotenv>=1.2.1",
]
